+++ src/main/scala/nl/knaw/dans/easy/license/FileAccessRight.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.license.internal
+package nl.knaw.dans.easy.license
 
 // TODO replace this object with a 'commons-library-call' (see also EASY-Stage-FileItem)
 object FileAccessRight extends Enumeration {
