+++ src/main/scala/nl/knaw/dans/easy/license/app/CommandLineOptions.scala
@@ -20,7 +20,8 @@ import javax.naming.Context
 import javax.naming.ldap.InitialLdapContext
 
 import com.yourmediashelf.fedora.client.{FedoraClient, FedoraCredentials}
-import nl.knaw.dans.easy.license.internal.{DatasetID, Parameters, Version}
+import nl.knaw.dans.easy.license.DatasetID
+import nl.knaw.dans.easy.license.internal.{Parameters, Version}
 import org.apache.commons.configuration.PropertiesConfiguration
 import org.rogach.scallop._
 import org.slf4j.LoggerFactory
