+++ src/main/scala/nl/knaw/dans/easy/license/internal/DatasetLoader.scala
@@ -17,7 +17,7 @@ package nl.knaw.dans.easy.license.internal
 
 import javax.naming.directory.Attributes
 
-import nl.knaw.dans.easy.license.internal.FileAccessRight.FileAccessRight
+import nl.knaw.dans.easy.license.{DatasetID, DepositorID, FileAccessRight, FileItem}
 import nl.knaw.dans.pf.language.emd.binding.EmdUnmarshaller
 import nl.knaw.dans.pf.language.emd.{EasyMetadata, EasyMetadataImpl, EmdAudience}
 import rx.lang.scala.schedulers.IOScheduler
@@ -64,8 +64,6 @@ case class Dataset(datasetID: DatasetID,
 
   def validate: Dataset = DatasetValidator.validate(this)
 }
-
-case class FileItem(path: String, accessibleTo: FileAccessRight, checkSum: String)
 
 trait DatasetLoader {
